refactor(threadpool): 统一线程池创建以符合云顶规范
- 审计模块使用ExecutorUtils替换ThreadPoolTaskExecutor创建线程池
- 缓存模块定时和专用线程池改用ExecutorUtils创建，简化代码并统一规范
- 搜索模块兜底线程池改为InternalExecutorFactory创建，适配DelegatingTaskExecutor
- 清理冗余线程工厂及注释，提升代码一致性和可维护性
- 资源关闭和拒绝策略配置与云顶规范保持一致

+++ ydsz-workflow/ydsz-workflow-server/src/main/java/com/njydsz/workflow/server/service/impl/FlowChartServiceImpl.java
@@ -33,15 +33,15 @@ import com.njydsz.workflow.server.service.FlowChartService;
 public class FlowChartServiceImpl implements FlowChartService {
 
   /** 节点默认宽度 */
-  private static final double NODE_WIDTH = 120;
+  private static final int NODE_WIDTH = 120;
   /** 节点默认高度 */
-  private static final double NODE_HEIGHT = 50;
+  private static final int NODE_HEIGHT = 50;
   /** 节点间距（水平） */
-  private static final double NODE_GAP_X = 60;
+  private static final int NODE_GAP_X = 60;
   /** 节点间距（垂直） */
-  private static final double NODE_GAP_Y = 80;
+  private static final int NODE_GAP_Y = 80;
   /** 绘图边距 */
-  private static final double PADDING = 40;
+  private static final int PADDING = 40;
 
   /** 自动布局：每行节点数 */
   private static final int NODES_PER_ROW = 3;
